model/requirements.txt
tensorflow
opencv-python
numpy
matplotlib
fastapi 
uvicorn 
python-multipart
pillow